feat(swap): use smart slippage (#4649)
