+++ apps/cowswap-frontend/src/modules/swap/hooks/useIsSlippageModified.ts
@@ -1,0 +1,7 @@
+import { useAtomValue } from 'jotai/index'
+
+import { slippageValueAndTypeAtom } from '../state/slippageValueAndTypeAtom'
+
+export function useIsSlippageModified() {
+  return useAtomValue(slippageValueAndTypeAtom).type === 'user'
+}
