+++ apps/cowswap-frontend/src/modules/swap/hooks/useIsSmartSlippageApplied.ts
@@ -1,0 +1,7 @@
+import { useAtomValue } from 'jotai/index'
+
+import { slippageValueAndTypeAtom } from '../state/slippageValueAndTypeAtom'
+
+export function useIsSmartSlippageApplied() {
+  return useAtomValue(slippageValueAndTypeAtom).type === 'smart'
+}
