+++ apps/cowswap-frontend/src/modules/swap/hooks/useSetSlippage.ts
@@ -1,0 +1,7 @@
+import { useSetAtom } from 'jotai'
+
+import { setSwapSlippageAtom } from '../state/slippageValueAndTypeAtom'
+
+export function useSetSlippage() {
+  return useSetAtom(setSwapSlippageAtom)
+}
