+++ apps/cowswap-frontend/src/modules/swap/state/EthFlow/updaters/index.ts
@@ -1,2 +1,1 @@
 export * from './EthFlowDeadlineUpdater'
-export * from './EthFlowSlippageUpdater'
