+++ apps/cowswap-frontend/src/modules/trade/index.ts
@@ -19,12 +19,14 @@ export * from './hooks/useSwitchTokensPlaces'
 export * from './hooks/useUpdateCurrencyAmount'
 export * from './hooks/useTradeRouteContext'
 export * from './hooks/useTradeConfirmState'
+export * from './hooks/useIsWrapOrUnwrap'
 export * from './containers/TradeWidget/types'
 export * from './utils/getReceiveAmountInfo'
 export * from './utils/parameterizeTradeRoute'
 export * from './state/receiveAmountInfoAtom'
 export * from './state/tradeTypeAtom'
 export * from './state/derivedTradeStateAtom'
+export * from './state/isWrapOrUnwrapAtom'
 export * from './pure/RecipientRow'
 export * from './pure/ReceiveAmountTitle'
 export * from './pure/PartnerFeeRow'
