+++ apps/cowswap-frontend/src/pages/Swap/index.tsx
@@ -6,7 +6,7 @@ import { useWalletInfo } from '@cowprotocol/wallet'
 import { Navigate, useLocation, useParams } from 'react-router-dom'
 
 import { AppDataUpdater } from 'modules/appData'
-import { SwapWidget, SwapDerivedStateUpdater, SwapAmountsFromUrlUpdater } from 'modules/swap'
+import { SwapWidget, SwapDerivedStateUpdater, SwapAmountsFromUrlUpdater, SmartSlippageUpdater } from 'modules/swap'
 import { useSwapSlippage } from 'modules/swap/hooks/useSwapSlippage'
 import { getDefaultTradeRawState } from 'modules/trade/types/TradeRawState'
 import { parameterizeTradeRoute } from 'modules/trade/utils/parameterizeTradeRoute'
@@ -30,6 +30,7 @@ export function SwapPage({ hooksEnabled }: SwapPageProps) {
       <AppDataUpdater orderClass="market" slippage={slippage} />
       <SwapDerivedStateUpdater />
       <SwapAmountsFromUrlUpdater />
+      <SmartSlippageUpdater />
       <SwapWidget hooksEnabled={hooksEnabled} />
     </>
   )
