+++ libs/common-const/src/bff.ts
@@ -1,1 +1,1 @@
-export const BFF_BASE_URL = process.env.REACT_APP_BFF_BASE_URL || 'https://bff.cow.fi'
+export const BFF_BASE_URL = process.env.REACT_APP_BFF_BASE_URL || 'https://bff.barn.cow.fi'
