+++ libs/common-utils/src/index.ts
@@ -55,3 +55,4 @@ export * from './tryParseCurrencyAmount'
 export * from './tryParseFractionalAmount'
 export * from './uriToHttp'
 export * from './userAgent'
+export * from './getCurrencyAddress'
